requirements.txt
asgiref==3.5.2
Django==4.0.5
django-ckeditor==6.4.2
django-js-asset==2.0.0
django-quill==0.1.0
Pillow==9.1.1
psycopg2-binary==2.9.3
PyMySQL==1.0.2
python-dotenv==0.20.0
sqlparse==0.4.2
whitenoise==6.2.0
